amplifier-bundle-orchestration/modules/amplifier-module-orchestrator-multi-context/pyproject.toml
[project]
name = "amplifier-orchestrator-multi-context"
version = "0.1.0"
description = "Multi-context workflow orchestrator for Amplifier Core"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "B", "A", "C4", "T10", "T20"]
ignore = ["ANN101", "ANN102"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
